Enhance tile modification handling: add MODIFY_TILE_LIST packet logging, refactor imagePath to use Optional, and update tile modification records for improved data handling

+++ starbound-protocol/src/main/java/irden/space/proxy/protocol/payload/packet/modify_tile_list/tile_modification/PlaceLiquid.java
@@ -1,7 +1,7 @@
 package irden.space.proxy.protocol.payload.packet.modify_tile_list.tile_modification;
 
 public record PlaceLiquid(
-        short liquidId,
+        int liquidId,
         float liquidLevel
 ) implements TileModification{
 }
